feature: #304 restructure part 1
